Add test when manager is stopped

+++ tests/system/test_fim/test_synchronization/test_synchronization.py
@@ -41,11 +41,14 @@ tags:
 """
 
 import os
+import json
 import pytest
+from time import sleep
+
 
 from wazuh_testing.tools.monitoring import HostMonitor
 from wazuh_testing.tools.system import HostManager
-from test_fim import create_folder_file, clean_logs
+from test_fim import create_folder_file, clean_logs, query_db
 
 
 # Hosts
@@ -58,18 +61,14 @@ local_path = os.path.dirname(os.path.abspath(__file__))
 messages_path = [os.path.join(local_path, 'data/messages.yml'),
                  os.path.join(local_path, 'data/delete_message.yml'),
                  os.path.join(local_path, 'data/wait_fim_scan.yml'),
-                ]
+                 os.path.join(local_path, 'data/agent_initializing_synchronization.yml'),
+                 os.path.join(local_path, 'data/mannager_initializing_synchronization.yml')
+                 ]
 tmp_path = os.path.join(local_path, 'tmp')
 scheduled_mode = 'testdir1'
+db_path = '/var/ossec/queue/db/001.db'
+db_script = '/var/system_query_db.py'
 
-
-
-def check_db(db_path, query):
-    #db_path = '/var/ossec/queue/db/001.db'
-    #query = " select * from fim_entry where full_path={}".format("/testdir1/testdir12.txt")
-    result = host_manager.run_command('wazuh-manager', "python /var/system_query_db.py {} '\"{}\"'".format(db_path, query))
-    
-    
 
 @pytest.mark.parametrize('case', ['add', 'modify', 'delete'])
 @pytest.mark.parametrize('folder_path', ['testdir1'])
@@ -81,7 +80,7 @@ def test_Synchronization_create_file_agent_stopped(folder_path, case):
     '''
     # Clear logs, create folder to monitored and restart the service
     clean_logs(host_manager)
-    create_folder_file(host_manager,folder_path)
+    create_folder_file(host_manager, folder_path)
     host_manager.control_service(host='wazuh-agent1', service='wazuh', state="restarted")
 
     # Check that the manager contains data to monitor
@@ -98,21 +97,82 @@ def test_Synchronization_create_file_agent_stopped(folder_path, case):
     host_manager.control_service(host='wazuh-agent1', service='wazuh', state="stopped")
 
     if (case == 'add'):
-        host_manager.run_command('wazuh-agent1', f'touch {folder_path}/{folder_path}2.txt')
+        host_manager.run_command('wazuh-agent1', f'touch {folder_path}/{folder_path}.txt')
 
     elif(case == 'modify'):
         host_manager.modify_file_content(host='wazuh-agent1', path=folder_path, content=folder_path)
 
     else:
         host_manager.run_command('wazuh-agent1', f'rm -rf {folder_path}')
-    
+        folder_path = f"'/{folder_path}/{folder_path}.txt'"
+        query = " select * from fim_entry where full_path='\"{}\"'".format(folder_path)
+
     # Start agent
     host_manager.control_service(host='wazuh-agent1', service='wazuh', state="started")
-    
 
     try:
-         HostMonitor(inventory_path=inventory_path,
-                     messages_path=messages_path[0],
-                     tmp_path=tmp_path).run()
+        HostMonitor(inventory_path=inventory_path,
+                    messages_path=messages_path[3],
+                    tmp_path=tmp_path).run()
+        if (case == 'delete'):
+            # Execute query to DB
+            sleep(5)
+            result = query_db(host_manager, db_script, db_path, f'\"{query}\"')
+            assert not json.loads(result)
+
     finally:
-         host_manager.run_command('wazuh-agent1', f'rm -rf {folder_path}')
+        host_manager.run_command('wazuh-agent1', f'rm -rf {folder_path}')
+
+
+@pytest.mark.parametrize('case', ['add', 'modify', 'delete'])
+@pytest.mark.parametrize('folder_path', ['testdir2'])
+def test_Synchronization_create_file_manager_stopped(folder_path, case):
+    '''
+    The test will monitor a directory and apply changes when agent is stopped.
+    Finally, it will verify that the FIM 'Synchronization' event is generated
+    in agent and manager side.
+    '''
+    # Clear logs, create folder to monitored and restart the service
+    clean_logs(host_manager)
+    create_folder_file(host_manager, folder_path)
+    host_manager.control_service(host='wazuh-agent1', service='wazuh', state="restarted")
+
+    # Check that the manager contains data to monitor
+    try:
+        HostMonitor(inventory_path=inventory_path,
+                    messages_path=messages_path[0],
+                    tmp_path=tmp_path).run()
+    except:
+        host_manager.run_command('wazuh-agent1', f'rm -rf {folder_path}')
+
+    clean_logs(host_manager)
+
+    # Stop mannager
+    host_manager.run_command('wazuh-manager', '/var/ossec/bin/wazuh-control stop')
+
+    if (case == 'add'):
+        host_manager.run_command('wazuh-agent1', f'touch {folder_path}/{folder_path}.txt')
+
+    elif(case == 'modify'):
+        host_manager.modify_file_content(host='wazuh-agent1', path=folder_path, content=folder_path)
+
+    else:
+        host_manager.run_command('wazuh-agent1', f'rm -rf {folder_path}')
+        folder_path = f"'/{folder_path}/{folder_path}.txt'"
+        query = " select * from fim_entry where full_path='\"{}\"'".format(folder_path)
+
+    # Start mannager
+    host_manager.control_service(host='wazuh-manager', service='wazuh', state="started")
+
+    try:
+        HostMonitor(inventory_path=inventory_path,
+                    messages_path=messages_path[4],
+                    tmp_path=tmp_path).run()
+        if (case == 'delete'):
+            # Execute query to DB
+            sleep(5)
+            result = query_db(host_manager, db_script, db_path, f'\"{query}\"')
+            assert not json.loads(result)
+
+    finally:
+        host_manager.run_command('wazuh-agent1', f'rm -rf {folder_path}')
